+++ src/layout/index.js
@@ -3,4 +3,3 @@
 require('./controllers');
 require('./states');
 require('./services');
-require('./directives');