+++ src/mockapi/index.js
@@ -9,21 +9,15 @@
 
 require('./providers');
 
-// TODO Not sure if there is a way, now that we are using CommonJS, to
-// eliminate this little IIFE.
-
-(function (mod) {
-  'use strict';
-
-  mod.run(function ($httpBackend, moondashMockRest) {
-
-    moondashMockRest.registerMocks($httpBackend);
+function ModuleRun ($httpBackend, MdMockRest) {
+    MdMockRest.registerMocks($httpBackend);
 
     // pass through everything else
     $httpBackend.whenGET(/\/*/).passThrough();
     $httpBackend.whenPOST(/\/*/).passThrough();
     $httpBackend.whenPUT(/\/*/).passThrough();
 
-  });
+}
 
-}(angular.module('moondashMock', ['moondash', 'ngMockE2E'])));
+angular.module('md.mockapi', [])
+  .run(ModuleRun);