+++ src/nav/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+// Define a submodule
+var angular = require('angular');
+angular.module('md.nav', []);
+
+
+require('./directives');
+require('./filters');