+++ test/protractor.conf.js
@@ -1,0 +1,19 @@
+// An example configuration file.
+config = require('../gulp/config');
+
+exports.config = {
+  seleniumServerJar: '../node_modules/protractor/selenium/selenium-server-standalone-2.44.0.jar',
+
+  // Spec patterns are relative to the current working directly when
+  // protractor is called.
+  specs: config.e2e.specs,
+
+  baseUrl: 'http://localhost:3000/',
+
+  // Options to be passed to Jasmine-node.
+  jasmineNodeOpts: {
+    showColors: true,
+    includeStackTrace: true,
+    defaultTimeoutInterval: 30000
+  }
+};